Utilisation de type int plutôt que float dans la classe Moving et ses descendants

+++ Frogger_processing/src/main/java/GameCommons/Game.java
@@ -206,7 +206,7 @@ public class Game implements IFrog, IEnvironment {
     }
 
     @Override
-    public void move_allCars(ArrayList<ArrayList<Car>> allCars, float ydir) {
+    public void move_allCars(ArrayList<ArrayList<Car>> allCars, int ydir) {
         for (ArrayList<Car> range_i : allCars){
             for (Car car : range_i){
                 car.move(0, ydir);
